requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
astroid==2.15.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
bleach==6.0.0
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cssselect==1.2.0
decorator==5.1.1
dicttoxml==1.7.16
dill==0.3.6
docutils==0.19
exceptiongroup==1.1.1
Flask==2.2.3
frozenlist==1.3.3
future==0.18.3
geocoder==1.38.1
geoip2==4.6.0
h11==0.14.0
idna==3.4
importlib-metadata==6.1.0
ip2geotools==0.1.6
IP2Location==8.9.0
isort==5.12.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
keyring==23.13.1
lazy-object-proxy==1.9.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
maxminddb==2.2.0
mccabe==0.7.0
mdurl==0.1.2
more-itertools==9.1.0
multidict==6.0.4
outcome==1.2.0
packaging==23.0
Pillow==9.4.0
pip-review==1.3.0
pkginfo==1.9.6
platformdirs==3.1.1
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
pylint==2.17.1
pyparsing==3.0.9
pyquery==2.0.0
PySocks==1.7.1
pywin32-ctypes==0.2.0
ratelim==0.1.6
readme-renderer==37.3
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.3.2
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
twine==4.0.2
typed-ast==1.5.4
typing==3.7.4.3
typing_extensions==4.5.0
urllib3==1.26.15
webencodings==0.5.1
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.15.0
wsproto==1.2.0
yarl==1.8.2
zipp==3.15.0
